backend/backend/requirements.txt
altgraph==0.17.4
anyio==4.9.0
asgiref==3.9.1
asttokens==2.2.1
async-timeout==5.0.1
attrs==23.1.0
autobahn==24.4.2
Automat==25.4.16
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.2.0
cli_helpers==2.4.0
click==8.1.3
colorama==0.4.5
comm==0.1.2
configobj==5.0.9
constantly==23.10.4
contourpy==1.1.0
cryptography==45.0.4
cycler==0.11.0
dacite==1.8.1
daphne==4.2.1
debugpy==1.6.4
decorator==5.1.1
Deprecated==1.2.14
distlib==0.4.0
distro==1.9.0
dj-database-url==2.1.0
Django==3.2.13
django-allauth==65.9.0
django-cors-headers==3.13.0
django-environ==0.8.1
django-heroku==0.3.1
Django-Verify-Email==3.0.3
djangorestframework==3.13.1
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
EasyProcess==1.1
entrypoint2==1.1
entrypoints==0.4
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==1.2.0
filelock==3.18.0
Flask==2.1.2
fonttools==4.42.0
future==0.18.3
graphtools==1.5.3
gunicorn==20.1.0
h11==0.16.0
heroku==0.1.4
htmlmin==0.1.12
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.4
ImageHash==4.3.1
incremental==24.7.2
inflection==0.5.1
ipykernel==6.19.4
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jiter==0.10.0
joblib==1.3.2
jupyter_client==7.4.8
jupyter_core==5.1.1
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
meld==1.0.2
meld3==2.0.1
MouseInfo==0.1.3
msgpack==1.1.1
mss==9.0.1
multimethod==1.9.1
mutagen==1.46.0
mycli==1.31.2
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nest-asyncio==1.5.6
networkx==3.1
npm==0.1.1
numpy==1.23.5
openai==1.94.0
opencv-python==4.8.1.78
openpyxl==3.1.5
optional-django==0.1.0
packaging==22.0
pandas==1.5.2
pandas-profiling==3.6.6
parso==0.8.3
patsy==0.5.3
pefile==2023.2.7
phik==0.12.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==4.3.8
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.7
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==1.10.12
pyfzf==0.3.1
pygame==2.3.0
PyGetWindow==0.0.9
Pygments==2.13.0
PyGSP==0.5.1
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.4.0
PyMsgBox==1.0.9
PyMySQL==1.1.1
pyOpenSSL==25.1.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
python-dotenv==1.1.1
pytweening==1.0.7
pytz==2022.7
PyWavelets==1.4.1
pywin32==305
pywin32-ctypes==0.2.3
PyYAML==6.0.1
pyzmq==24.0.1
redis==6.2.0
requests==2.31.0
rich==13.5.2
scikit-learn==1.3.0
scipy==1.11.1
scprep==1.2.3
seaborn==0.12.2
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.4.1
split-folders==0.5.1
sqlglot==26.30.0
sqlparse==0.4.3
stack-data==0.6.2
statsmodels==0.14.0
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tasklogger==1.2.0
threadpoolctl==3.2.0
tomli==2.2.1
topping==0.1.1
tornado==6.2
tqdm==4.66.1
traitlets==5.8.0
ttkthemes==3.2.2
Twisted==25.5.0
txaio==25.6.1
typeguard==2.13.3
typing_extensions==4.14.1
tzdata==2022.6
uritemplate==4.2.0
urllib3==2.0.4
virtualenv==20.32.0
visions==0.7.5
wcwidth==0.2.5
Werkzeug==2.1.2
whitenoise==6.5.0
wordcloud==1.9.2
wordcount==1.0
wrapt==1.15.0
ydata-profiling==4.5.0
zope.interface==7.2
